删除话题

+++ app/http/controllers/api/v1/topics_controller.go
@@ -90,6 +90,12 @@ func (ctrl *TopicsController) Delete(c *gin.Context) {
 		response.Abort404(c)
 		return
 	}
+
+	if ok := policies.CanModifyTopic(c, topicModel); !ok {
+		response.Abort403(c)
+		return
+	}
+
 	rowsAffected := topicModel.Delete()
 	if rowsAffected > 0 {
 		response.Success(c)
